Reorganises how we internally construct the webpack configuration 😬
