+++ src/webpack.config/index.spec.js
@@ -13,16 +13,6 @@ const baseSaguiConfig = {
 }
 
 describe('webpack', function () {
-  it('should allow disabling all loaders', () => {
-    const saguiConfig = {
-      ...baseSaguiConfig,
-      disabledLoaders: ['font', 'image', 'javaScript', 'json', 'style', 'txt', 'video', 'yaml']
-    }
-
-    const config = webpack(saguiConfig).webpack[0]
-    expect(config.module.loaders).to.be.undefined
-  })
-
   describe('webpackConfig extension', function () {
     it('should allow extending the default configuration', function () {
       const saguiConfig = {
@@ -32,13 +22,13 @@ describe('webpack', function () {
         }
       }
 
-      const config = webpack(saguiConfig).webpack[0]
+      const config = webpack(saguiConfig)[0]
 
       expect(config.target).equal('electron')
     })
 
     it('should allow overwriting the default configuration', function () {
-      const defaultConfig = webpack(baseSaguiConfig).webpack[0]
+      const defaultConfig = webpack(baseSaguiConfig)[0]
       expect(defaultConfig.devtool).equal('inline-source-map')
 
       const saguiConfig = {
@@ -48,7 +38,7 @@ describe('webpack', function () {
         }
       }
 
-      const config = webpack(saguiConfig).webpack[0]
+      const config = webpack(saguiConfig)[0]
       expect(config.devtool).equal('cheap-source-map')
     })
 
@@ -68,7 +58,7 @@ describe('webpack', function () {
         }
       }
 
-      const config = webpack(saguiConfig).webpack[0]
+      const config = webpack(saguiConfig)[0]
       const loaders = config.module.loaders.filter((loader) => loader.loader === 'babel')
 
       // should change the existing loader, not add another
