+++ src/webpack.config/loaders/font.js
