+++ src/webpack.config/loaders/image.js
