+++ src/webpack.config/loaders/image.spec.js
