+++ src/webpack.config/loaders/index.js
