+++ src/webpack.config/loaders/javascript.js
