+++ src/webpack.config/loaders/javascript.spec.js
