+++ src/webpack.config/loaders/json.js
