+++ src/webpack.config/loaders/json.spec.js
