+++ src/webpack.config/loaders/style.js
