+++ src/webpack.config/loaders/style.spec.js
