+++ src/webpack.config/loaders/txt.js
