+++ src/webpack.config/loaders/txt.spec.js
