+++ src/webpack.config/loaders/video.js
