+++ src/webpack.config/loaders/yaml.js
