--- src/webpack/presets/define-node-env.js
@@ -1,14 +1,0 @@
-import { DefinePlugin } from 'webpack'
-
-export default {
-  name: 'define-node-env',
-  configure () {
-    return {
-      plugins: [
-        new DefinePlugin({
-          'process.env.NODE_ENV': JSON.stringify(process.env.NODE_ENV)
-        })
-      ]
-    }
-  }
-}
